chore: Added types to module interface
