+++ .github/workflows/pr.yml
@@ -25,6 +25,9 @@ jobs:
         working-directory: package
         run: yarn build
 
+      - name: Typecheck
+        run: yarn workspaces foreach run ci:typecheck
+
       - name: Install Playwright deps
         run: yarn playwright install-deps chrome firefox webkit
