+++ package.json
@@ -17,11 +17,13 @@
     },
     "packageManager": "yarn@3.4.1",
     "devDependencies": {
+        "@types/node": "16",
         "@types/wait-on": "^5.3.1",
         "playwright": "^1.32.2",
         "prettier": "^2.8.4",
         "process": "^0.11.10",
         "ts-node": "^10.9.1",
+        "typescript": "^5.0.4",
         "wait-on": "^7.0.1"
     },
     "resolutions": {
