+++ package/package.json
@@ -28,7 +28,8 @@
     "files": [
         "release",
         "README.md",
-        "copyMonacoFilesAMD"
+        "copyMonacoFilesAMD",
+        "globalApi.d.ts"
     ],
     "bin": {
         "copyMonacoFilesAMD": "./copyMonacoFilesAMD.js"
@@ -59,7 +60,7 @@
         "rollup": "^3.17.3",
         "terser": "^5.16.2",
         "ts-node": "^10.9.1",
-        "typescript": "^4.0.0",
+        "typescript": "^5.0.4",
         "vscode-languageserver-textdocument": "1.0.4"
     },
     "dependencies": {
