+++ package/scripts/tsconfig.build.json
@@ -11,8 +11,6 @@
         "declaration": true,
         "noEmit": false,
         "types": [
-            "../monaco",
-            "monaco-editor/monaco",
             "@kusto/language-service/Kusto.JavaScript.Client",
             "@kusto/language-service-next/Kusto.Language.Bridge"
         ]
