+++ package/scripts/tsconfig.watch.json
@@ -1,7 +1,7 @@
 {
     "$schema": "https://json.schemastore.org/tsconfig",
     "extends": "../tsconfig.json",
-    "include": ["../src/monaco.contribution.ts", "../src/KustoWorker.ts", "../src/KustoMode.ts"],
+    "include": ["../src/**/*"],
     "compilerOptions": {
         "rootDir": "../src",
         "module": "esnext",
@@ -10,8 +10,6 @@
         "preserveWatchOutput": true,
         "pretty": true,
         "types": [
-            "../monaco",
-            "monaco-editor/monaco",
             "@kusto/language-service/Kusto.JavaScript.Client",
             "@kusto/language-service-next/Kusto.Language.Bridge"
         ]
