+++ package/src/languageService/kustoMonarchLanguageDefinition.ts
@@ -1,3 +1,5 @@
+import type * as monaco from 'monaco-editor/esm/vs/editor/editor.worker';
+
 export const KustoLanguageDefinition: monaco.languages.IMonarchLanguage = {
     name: 'kusto',
     mimeTypes: ['text/kusto'],
