+++ package/tsconfig.json
@@ -12,8 +12,6 @@
         "allowSyntheticDefaultImports": true,
         "isolatedModules": true,
         "types": [
-            "./monaco",
-            "monaco-editor/monaco",
             "@kusto/language-service/Kusto.JavaScript.Client",
             "@kusto/language-service-next/Kusto.Language.Bridge"
         ]
