+++ samples/amd-webpack-react/.babelrc
@@ -7,6 +7,7 @@
                 "modules": false
             }
         ],
-        "@babel/preset-react"
+        "@babel/preset-react",
+        "@babel/preset-typescript"
     ]
 }
