+++ samples/amd-webpack-react/package.json
@@ -11,7 +11,8 @@
         "serve": "webpack serve",
         "copy-amd-files": "copyMonacoFilesAMD ./public",
         "playwright:prepare": "yarn copy-amd-files",
-        "playwright:webserver": "webpack serve --port 3000 --no-open"
+        "playwright:webserver": "webpack serve --port 3000 --no-open",
+        "ci:typecheck": "tsc"
     },
     "dependencies": {
         "@kusto/monaco-kusto": "workspace:*",
@@ -23,10 +24,15 @@
         "@babel/core": "^7.21.4",
         "@babel/preset-env": "^7.21.4",
         "@babel/preset-react": "^7.18.6",
+        "@babel/preset-typescript": "^7.21.4",
+        "@tsconfig/create-react-app": "^1.0.3",
+        "@types/react": "^18.0.38",
+        "@types/react-dom": "^18.0.11",
         "babel-loader": "^9.1.2",
         "css-loader": "^6.7.3",
         "html-webpack-plugin": "^5.5.0",
         "style-loader": "^3.3.2",
+        "typescript": "^5.0.4",
         "webpack": "^5.78.0",
         "webpack-cli": "^5.0.1",
         "webpack-dev-server": "^4.13.2"
