+++ samples/amd-webpack-react/src/global.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@kusto/monaco-kusto/globalApi" />
