+++ samples/amd-webpack-react/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "@tsconfig/create-react-app/tsconfig.json",
+    "include": ["./src/**/*"]
+}
