+++ samples/parcel/package.json
@@ -6,21 +6,19 @@
     "scripts": {
         "start": "parcel index.html",
         "build": "parcel build index.html",
-        "playwright:webserver": "parcel index.html --port 3000"
+        "playwright:webserver": "parcel index.html --port 3000",
+        "ci:typecheck": "tsc"
     },
     "alias": {
         "process": false
     },
     "dependencies": {
         "@kusto/monaco-kusto": "workspace:*",
-        "monaco-editor": "~0.35.0",
-        "react": "^16.13.0",
-        "react-dom": "^16.13.1"
+        "monaco-editor": "~0.35.0"
     },
     "devDependencies": {
-        "@types/react": "^16.9.11",
-        "@types/react-dom": "^16.8.4",
+        "@tsconfig/recommended": "^1.0.2",
         "parcel": "^2.0.0",
-        "typescript": "^4.3.3"
+        "typescript": "^5.0.4"
     }
 }
