+++ samples/parcel/tsconfig.json
@@ -1,10 +1,9 @@
 {
     "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "@tsconfig/recommended",
     "compilerOptions": {
-        "jsx": "react",
-        "moduleResolution": "node",
-        "allowSyntheticDefaultImports": true,
-        "strict": true,
-        "isolatedModules": true
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "resolveJsonModule": true,
+        "noEmit": true
     }
 }
